backend/requirements.txt
flask==2.3.3
python-dotenv==1.0.0
psycopg2-binary>=2.9.0
anthropic>=0.18.0
PyPDF2>=3.0.0
pytesseract>=0.3.10
pillow>=10.0.0
python-docx>=0.8.11
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.0
pdf2image>=1.16.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
pytest>=7.0.0
pytest-flask>=1.2.0
beautifulsoup4>=4.12.0
email-validator>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
flask-cors==4.0.0
